requirements.txt
asgiref==3.6.0
attrs==22.2.0
boto3==1.26.94
botocore==1.29.94
dj-database-url==1.2.0
Django==4.1.7
django-filter==22.1
django-simple-history==3.3.0
django-softdelete==0.10.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
iniconfig==2.0.0
jmespath==1.0.1
Markdown==3.4.1
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.5
PyJWT==2.6.0
pytest==7.2.2
python-dateutil==2.8.2
pytz==2022.7.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.15
